The restructuring is complete, I have added links to the stage description. What I do not know yet how to add several positions, which would either require that each stage parameters are an N_p1 long array, or we add N_p1 parameter classes, which would be quite terrible to handle. I think, the sample position part can be deprecated for the sake of the NEXUS base class stage transformation.
The added symbols are used as internal variables thus they can occur in multiple places. We still need to find a way of defining the signal variable, which depends on how the measurement was done (what data is available). In general, it would be nice to add simply the first two dimensions of the data.

Now, I request a merge thus this can form the basis of our ellipsometry application definition.
I have tested the ..._with_NEXUS_fields.yaml file, it is parsed properly by pyaml.