Some more resturcturing of the class. The environmental parameters, such as:  - medium around the sample -- at least state air  - medium refractive index  - manipulation of the sample  - all methods: optical, electric, mechanical, pH, etc.
are now under an NXenvironment. This comes optionally with a Sensor field, which can specify all but the optical excitation of the game.
The optical part is under an NXsource, modified the same way as the source of the ellipsometer.

The resulted structure is simpler and easier to follow upon.